Make spacing functions work with variables and expressions
Previously, they've been only working with constant numbers because of the
`-distance` bit in the code. This change adds a `negative` helper function
that returns a negative of the correct type, depending on its parameter.

This also adds the possibility to use a getter in the spacing functions for the
sake of consistency.